refactor: move last will store in own folder
