+++ frontend/src/app/(dynamic)/last-will/editor/marriage/page.tsx
@@ -16,7 +16,7 @@ import { TextInput } from '../../../../../components/Form/TextInput/TextInput'
 import { Headline } from '../../../../../components/Headline/Headline'
 import { routes } from '../../../../../services/routes/routes'
 import { useAppDispatch, useAppSelector } from '../../../../../store/hooks'
-import { sendLastWillState, setMarriage, setProgressKeys } from '../../../../../store/lastwill'
+import { sendLastWillState, setMarriage, setProgressKeys } from '../../../../../store/lastwill/lastwill'
 import { Gender } from '../../../../../types/gender'
 import { MarriageFormPayload, Person } from '../../../../../types/lastWill'
 import { SidebarPages } from '../../../../../types/sidebar'
