+++ frontend/src/app/(dynamic)/last-will/editor/testator/page.tsx
@@ -14,7 +14,7 @@ import { TextInput } from '../../../../../components/Form/TextInput/TextInput'
 import { Headline } from '../../../../../components/Headline/Headline'
 import { routes } from '../../../../../services/routes/routes'
 import { useAppDispatch, useAppSelector } from '../../../../../store/hooks'
-import { sendLastWillState, setProgressKeys, setTestator } from '../../../../../store/lastwill'
+import { sendLastWillState, setProgressKeys, setTestator } from '../../../../../store/lastwill/lastwill'
 import { Gender } from '../../../../../types/gender'
 import { TestatorFormPayload } from '../../../../../types/lastWill'
 import { SidebarPages } from '../../../../../types/sidebar'
