+++ frontend/src/components/Modal/HeirsModal/HeirsOrganisationModal/HeirsOrganisationModal.tsx
@@ -2,7 +2,7 @@ import { nanoid } from '@reduxjs/toolkit'
 import { Form, Formik } from 'formik'
 import { ObjectSchema, object, string } from 'yup'
 import { useAppDispatch, useAppSelector } from '../../../../store/hooks'
-import { addOrganisationHeir, sendLastWillState, updateOrganisationHeir } from '../../../../store/lastwill'
+import { addOrganisationHeir, sendLastWillState, updateOrganisationHeir } from '../../../../store/lastwill/lastwill'
 import { Organisation, OrganisationFormPayload } from '../../../../types/lastWill'
 import { Button } from '../../../ButtonsAndLinks/Button/Button'
 import { TextInput } from '../../../Form/TextInput/TextInput'
