+++ frontend/src/components/Modal/HeirsModal/HeirsPersonModal/HeirsPersonModal.tsx
@@ -5,7 +5,7 @@ import { personMoreInfosOptions } from '../../../../../content/checkboxOptions'
 import { childRelationshipOptions, genderOptions, heirsPersonType } from '../../../../../content/dropdownOptions'
 import { heirsTypes } from '../../../../services/heirs'
 import { useAppDispatch, useAppSelector } from '../../../../store/hooks'
-import { addPersonHeir, sendLastWillState, updatePersonHeir } from '../../../../store/lastwill'
+import { addPersonHeir, sendLastWillState, updatePersonHeir } from '../../../../store/lastwill/lastwill'
 import { Gender } from '../../../../types/gender'
 import { ChildRelationShip, HeirsTypes, Person, PersonFormPayload, PersonType } from '../../../../types/lastWill'
 import { Button } from '../../../ButtonsAndLinks/Button/Button'
